# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Blossomer GTM API v2",
...
    "contact": {
      "name": "Blossomer GTM API Support",
      "email": "support@blossomer.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    },
    {
      "url": "https://api.blossomer.com",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/api/campaigns/positioning": {
      "post": {
        "tags": [
          "Campaigns",
          "Campaigns",
          "Positioning",
          "AI"
        ],
        "summary": "Generate Unique Insight and Unique Selling Points",
        "operationId": "generate_positioning_api_campaigns_positioning_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Data"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "A unique insight and a list of unique selling points for the given company context.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PositioningResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/campaigns/generate-email": {
      "post": {
        "tags": [
          "Campaigns",
          "Campaigns",
          "Email Generation",
          "AI"
        ],
        "summary": "Generate Email Campaign",
        "description": "Generate a personalized email campaign based on company context, target account/persona,\nand user preferences from the Email Campaign Wizard.\n\nThis endpoint synthesizes:\n- Company overview and capabilities\n- Target account firmographics and buying signals\n- Target persona demographics and use cases\n- User preferences for emphasis, opening line strategy, and CTA type\n\nReturns structured email content with modular segments and metadata for UI rendering.",
        "operationId": "generate_email_api_campaigns_generate_email_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailGenerationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "A complete email campaign with subjects, body segments, and breakdown metadata.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailGenerationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/campaigns/campaigns-crud": {
      "post": {
        "tags": [
          "Campaigns"
        ],
        "summary": "Create Campaign",
        "description": "Create a new campaign for an account and persona.\n\n- **account_id**: Account ID (must be owned by authenticated user via company)\n- **persona_id**: Persona ID (must belong to the account)\n- **name**: Campaign name (required, max 255 chars)\n- **campaign_type**: Campaign type (email, linkedin, cold_call, ad)\n- **campaign_data**: JSON data with subject_line, content, segments, etc.",
        "operationId": "create_campaign_api_campaigns_campaigns_crud_post",
        "parameters": [
          {
            "name": "account_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Account ID to create campaign for",
              "title": "Account Id"
            },
            "description": "Account ID to create campaign for"
          },
          {
            "name": "persona_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Persona ID to create campaign for",
              "title": "Persona Id"
            },
            "description": "Persona ID to create campaign for"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CampaignCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Campaigns"
        ],
        "summary": "Get Campaigns",
        "description": "Get all campaigns for an account.\n\n- **account_id**: Account ID (must be owned by authenticated user via company)\n- **persona_id**: Optional persona ID to filter campaigns\n- **skip**: Number of campaigns to skip (for pagination)\n- **limit**: Maximum number of campaigns to return (1-1000)",
        "operationId": "get_campaigns_api_campaigns_campaigns_crud_get",
        "parameters": [
          {
            "name": "account_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Account ID to get campaigns for",
              "title": "Account Id"
            },
            "description": "Account ID to get campaigns for"
          },
          {
            "name": "persona_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional persona ID to filter by",
              "title": "Persona Id"
            },
            "description": "Optional persona ID to filter by"
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of campaigns to skip",
              "default": 0,
              "title": "Skip"
            },
            "description": "Number of campaigns to skip"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "description": "Maximum number of campaigns to return",
              "default": 100,
              "title": "Limit"
            },
            "description": "Maximum number of campaigns to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CampaignResponse"
                  },
                  "title": "Response Get Campaigns Api Campaigns Campaigns Crud Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/campaigns/campaigns-crud/{campaign_id}": {
      "get": {
        "tags": [
          "Campaigns"
        ],
        "summary": "Get Campaign",
        "description": "Get a specific campaign by ID.\n\nOnly returns campaigns owned by the authenticated user (via account->company).",
        "operationId": "get_campaign_api_campaigns_campaigns_crud__campaign_id__get",
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Campaign Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Campaigns"
        ],
        "summary": "Update Campaign",
        "description": "Update a campaign.\n\nOnly updates campaigns owned by the authenticated user (via account->company).\nAll fields are optional - only provided fields will be updated.",
        "operationId": "update_campaign_api_campaigns_campaigns_crud__campaign_id__put",
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Campaign Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CampaignUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Campaigns"
        ],
        "summary": "Delete Campaign",
        "description": "Delete a campaign.\n\nOnly deletes campaigns owned by the authenticated user (via account->company).",
        "operationId": "delete_campaign_api_campaigns_campaigns_crud__campaign_id__delete",
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Campaign Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/demo/campaigns/generate-email": {
      "post": {
        "tags": [
          "Demo",
          "Demo",
          "Campaigns",
          "Email Generation",
          "AI"
        ],
        "summary": "[DEMO] Generate Email Campaign",
        "description": "Generate a personalized email campaign for demo users,\nwith IP-based rate limiting.",
        "operationId": "demo_generate_email_demo_campaigns_generate_email_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailGenerationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "A complete email campaign with subjects, body segments, and breakdown metadata.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailGenerationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CampaignCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "title": "Name"
          },
          "campaign_type": {
            "type": "string",
            "maxLength": 50,
            "title": "Campaign Type",
            "description": "Campaign type: email, linkedin, cold_call, ad"
          },
          "campaign_data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Campaign Data",
            "description": "Campaign data including subject_line, content, segments, alternatives, configuration"
          }
        },
        "type": "object",
        "required": [
          "name",
          "campaign_type",
          "campaign_data"
        ],
        "title": "CampaignCreate",
        "description": "Schema for creating a new campaign.",
        "example": {
          "campaign_data": {
            "alternatives": {
              "subject_lines": [
                "Scaling your engineering team at {{company}}?",
                "How {{company}} can ship code 40% faster"
              ]
            },
            "configuration": {
              "length": "short",
              "personalization": "high",
              "tone": "professional"
            },
            "content": "Hi {{name}}, I noticed {{company}} has been growing rapidly...",
            "segments": [
              {
                "text": "Hi {{name}}",
                "type": "greeting"
              },
              {
                "text": "I noticed {{company}} has been growing rapidly and hiring more developers",
                "type": "opening"
              },
              {
                "text": "As teams scale, manual code reviews and slow CI/CD pipelines often become major bottlenecks",
                "type": "pain-point"
              },
              {
                "text": "TechFlow's AI-powered automation platform helps engineering teams like yours maintain velocity while improving code quality",
                "type": "solution"
              },
              {
                "text": "We've helped 50+ similar companies reduce their build times by 40% and speed up code reviews by 60%",
                "type": "evidence"
              },
              {
                "text": "Would you be open to a 15-minute demo to see how this could work for your team?",
                "type": "cta"
              }
            ],
            "subject_line": "Quick question about your development workflow"
          },
          "campaign_type": "email",
          "name": "Q4 VP Engineering Outreach"
        }
      },
      "CampaignResponse": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "title": "Name"
          },
          "campaign_type": {
            "type": "string",
            "maxLength": 50,
            "title": "Campaign Type",
            "description": "Campaign type: email, linkedin, cold_call, ad"
          },
          "campaign_data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Campaign Data",
            "description": "Campaign data including subject_line, content, segments, alternatives, configuration"
          },
          "id": {
            "type": "string",
            "format": "uuid4",
            "title": "Id"
          },
          "account_id": {
            "type": "string",
            "format": "uuid4",
            "title": "Account Id"
          },
          "persona_id": {
            "type": "string",
            "format": "uuid4",
            "title": "Persona Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "name",
          "campaign_type",
          "campaign_data",
          "id",
          "account_id",
          "persona_id",
          "created_at",
          "updated_at"
        ],
        "title": "CampaignResponse",
        "description": "Schema for campaign responses."
      },
      "CampaignUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "campaign_type": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Campaign Type"
          },
          "campaign_data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Campaign Data"
          }
        },
        "type": "object",
        "title": "CampaignUpdate",
        "description": "Schema for updating campaign information."
      },
...
      "EmailGenerationMetadata": {
        "properties": {
          "generation_id": {
            "type": "string",
            "title": "Generation Id",
            "description": "Unique identifier for this generation"
          },
          "confidence": {
            "type": "string",
            "title": "Confidence",
            "description": "Confidence level for email quality: high|medium|low"
          },
          "personalization_level": {
            "type": "string",
            "title": "Personalization Level",
            "description": "Level of personalization achieved: high|medium|low"
          },
          "processing_time_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Processing Time Ms",
            "description": "Time taken to generate email in milliseconds"
          }
        },
        "type": "object",
        "required": [
          "generation_id",
          "confidence",
          "personalization_level"
        ],
        "title": "EmailGenerationMetadata",
        "description": "Metadata about the email generation process."
      },
      "EmailGenerationRequest": {
        "properties": {
          "company_context": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Context",
            "description": "Company overview from localStorage dashboard_overview"
          },
          "target_account": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Account",
            "description": "Selected target account from wizard step 1"
          },
          "target_persona": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Persona",
            "description": "Selected target persona from wizard step 1"
          },
          "preferences": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferences",
            "description": "User preferences from wizard steps 2-3"
          }
        },
        "type": "object",
        "title": "EmailGenerationRequest",
        "description": "Request model for email generation API."
      },
      "EmailGenerationResponse": {
        "properties": {
          "subjects": {
            "$ref": "#/components/schemas/EmailSubjects",
            "description": "Generated subject lines"
          },
          "email_body": {
            "items": {
              "$ref": "#/components/schemas/EmailSegment"
            },
            "type": "array",
            "title": "Email Body",
            "description": "Email content broken into structured segments"
          },
          "breakdown": {
            "additionalProperties": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            },
            "type": "object",
            "title": "Breakdown",
            "description": "Flexible segment breakdown for UI rendering - maps segment types to {label, description, color}"
          },
          "metadata": {
            "$ref": "#/components/schemas/EmailGenerationMetadata",
            "description": "Generation metadata and quality metrics"
          }
        },
        "type": "object",
        "required": [
          "subjects",
          "email_body",
          "breakdown",
          "metadata"
        ],
        "title": "EmailGenerationResponse",
        "description": "Response model for email generation API."
      },
      "EmailSegment": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text",
            "description": "The text content of this email segment"
          },
          "type": {
            "type": "string",
            "title": "Type",
            "description": "Segment type: greeting|opening|pain-point|solution|evidence|cta|signature"
          }
        },
        "type": "object",
        "required": [
          "text",
          "type"
        ],
        "title": "EmailSegment",
        "description": "Individual email segment with type classification."
      },
      "EmailSubjects": {
        "properties": {
          "primary": {
            "type": "string",
            "title": "Primary",
            "description": "Most effective subject line"
          },
          "alternatives": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 2,
            "minItems": 2,
            "title": "Alternatives",
            "description": "2 alternative subject lines"
          }
        },
        "type": "object",
        "required": [
          "primary",
          "alternatives"
        ],
        "title": "EmailSubjects",
        "description": "Generated email subjects with primary and alternatives."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PositioningResponse": {
        "properties": {
          "unique_insight": {
            "type": "string",
            "title": "Unique Insight",
            "description": "Core reframe of the problem (unique insight)"
          },
          "unique_selling_points": {
            "items": {
              "$ref": "#/components/schemas/UniqueSellingPoint"
            },
            "type": "array",
            "title": "Unique Selling Points",
            "description": "List of unique selling points (USPs)"
          }
        },
        "type": "object",
        "required": [
          "unique_insight",
          "unique_selling_points"
        ],
        "title": "PositioningResponse"
      },
...
      "UniqueSellingPoint": {
        "properties": {
          "theme": {
            "type": "string",
            "title": "Theme",
            "description": "USP name or theme"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "description": "Brief benefit or outcome"
          },
          "evidence": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Evidence",
            "description": "Supporting details or explanations"
          }
        },
        "type": "object",
        "required": [
          "theme",
          "description",
          "evidence"
        ],
        "title": "UniqueSellingPoint"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}